debug minor items

+++ install.sh
@@ -2,14 +2,6 @@
 echo "Installer for PASTA database on Debian/Ubuntu-Linux"
 echo "IMPORTANT: if you have problems, visit https://jugit.fz-juelich.de/pasta/main/-/wikis/home#installation-scripts"
 echo "Default choices are accepted by return: [Y/n]->yes; [default]->default"
-echo
-read -p "Do you wish to start installing now [Y/n] ? " yesno
-if [[ $yesno = 'N' ]] || [[ $yesno = 'n' ]]
-then
-  echo "  Did not install anything"
-  exit
-fi
-echo
 
 echo "Ensure installer has sudo rights"
 if [ "$EUID" -ne 0 ]
@@ -19,7 +11,6 @@ if [ "$EUID" -ne 0 ]
 fi
 echo
 THEUSER=$(logname)
-
 echo "Ensure that conda is not running"
 if conda &>/dev/null
 then
@@ -29,9 +20,46 @@ else
   echo "  Info: 'conda' is not installed."
 fi
 echo
+######## ASK QUESTIONS #############
+OUTPUT=$(sudo -u $THEUSER git config -l | grep "user")
+if [[ -n $OUTPUT ]]
+then
+  echo "  git user and email are set"
+else
+  echo "  Set your git user information"
+  read -p "  What is your name? " GIT_NAME
+  read -p "  What is your email? " GIT_EMAIL
+fi
+echo
 
+echo "Two empty (for safety) directories are required. One for the source code"
+echo "and the other as central place to store data, work in."
+read -p "  Where to store the source code? [pastaSource, i.e. /home/${THEUSER}/pastaSource] " pasta_src
+read -p "  Where to store the data? [pastaData, i.e. /home/${THEUSER}/pastaData] " pasta
+if [ -z $pasta_src ]
+then
+  pasta_src="pasta_source"
+fi
+if [ -z $pasta ]
+then
+  pasta="pasta"
+fi
+echo
+####################################
+
+echo
+read -p "Do you wish to install everything [Y/n] ? " yesno
+if [[ $yesno = 'N' ]] || [[ $yesno = 'n' ]]
+then
+  echo "  Did not install anything"
+  exit
+fi
+echo
 sudo rm installPASTA2.log
+sudo -u $THEUSER mkdir /home/$THEUSER/$pasta
 
+
+# #########  START INSTALLATION  #################
 echo "Ensure python, pip and pandoc are installed. This takes a few minutes"
 sudo add-apt-repository -y universe                    >> installPASTA2.log
 sudo apt-get install -y python3 python3-pip pandoc npm >> installPASTA2.log
@@ -50,14 +78,58 @@ if [[ -n $OUTPUT ]]
 then
   echo "  git user and email are set"
 else
-  echo "  Set your git user information"
-  read -p "  What is your name? " GIT_NAME
-  read -p "  What is your email? " GIT_EMAIL
   sudo -u $THEUSER git config --global --add user.name "${GIT_NAME}"
   sudo -u $THEUSER git config --global --add user.email $GIT_EMAIL
 fi
 echo
 
+echo "Start cloning the git repositories: tools, python-backend, javascript-frontend"
+sudo -u $THEUSER git clone git@github.com:PASTA-ELN/desktop.git $pasta_src >> installPASTA2.log  2>&1
+echo
+
+echo "Adopt path and python-path in your environment"
+sudo -u $THEUSER echo "#PASTA changes" >> /home/$THEUSER/.bashrc
+sudo -u $THEUSER echo "export PATH=\$PATH:/home/${THEUSER}/${pasta_src}/Python" >> /home/$THEUSER/.bashrc
+sudo -u $THEUSER echo "export PYTHONPATH=\$PYTHONPATH:/home/${THEUSER}/${pasta_src}/Python" >> /home/$THEUSER/.bashrc
+echo
+
+echo "Install python requirements. This takes a few minutes."
+cd /home/$THEUSER/$pasta_src/main
+sudo -H pip3 install -r requirements.txt           >> installPASTA2.log
+echo
+
+echo "Create PASTA configuration file .pasta.json in home directory"
+
+# sudo -u $THEUSER echo "{" > /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"softwareDir\": \"/home/${THEUSER}/${pasta_src}/main\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"userID\": \"${THEUSER}\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"version\": 1," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"extractors\": {}," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"qrPrinter\": {}," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"default\": \"research\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"magicTags\": [\"P1\",\"P2\",\"P3\",\"TODO\",\"WAIT\",\"DONE\"]," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  " >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"links\": {" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    \"research\": {" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "      \"local\": {" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "        \"user\": \"${CDB_USER}\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "        \"password\": \"${CDB_PASSW}\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "        \"database\": \"pasta_tutorial\"," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "        \"path\": \"/home/${THEUSER}/${pasta}\"" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "      }," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "      \"remote\": {}" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    }" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  }," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  " >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  \"tableFormat\": {" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    \"x0\":{\"-label-\":\"Projects\",\"-default-\": [22,6,50,22]}," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    \"measurement\":{\"-default-\": [24,7,23,23,-5,-6,-6,-6]}," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    \"sample\":{\"-default-\": [23,23,23,23,-5]}," >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "    \"procedure\":{\"-default-\": [20,20,20,40]}" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "  }" >> /home/$THEUSER/.pastaELN.json
+# sudo -u $THEUSER echo "}" >> /home/$THEUSER/.pastaELN.json
+# sudo chown $THEUSER:$THEUSER /home/$THEUSER/.pastaELN.json
+echo
 
 CDB_USER="admin"
 CDB_PASSW=$(date +%s | sha256sum | base64 | head -c 12 ; echo)
@@ -77,77 +149,6 @@ else
   curl -X PUT http://$CDB_USER:$CDB_PASSW@127.0.0.1:5984/_replicator      >> installPASTA2.log  2>&1
   curl -X PUT http://$CDB_USER:$CDB_PASSW@127.0.0.1:5984/_global_changes  >> installPASTA2.log  2>&1
 fi
-echo "  WRITE DOWN THIS PASSWORD FOR SAFEKEEPING: $CDB_PASSW"
-echo
-
-
-echo "Two empty (for safety) directories are required. One for the source code"
-echo "and the other as central place to store data, work in."
-read -p "  Where to store the source code? [pasta_source, i.e. /home/${THEUSER}/pasta_source] " pasta_src
-read -p "  Where to store the data? [pasta, i.e. /home/${THEUSER}/pasta] " pasta
-if [ -z $pasta_src ]
-then
-  pasta_src="pasta_source"
-fi
-if [ -z $pasta ]
-then
-  pasta="pasta"
-fi
-sudo -u $THEUSER mkdir /home/$THEUSER/$pasta_src
-sudo -u $THEUSER mkdir /home/$THEUSER/$pasta
-echo
-
-
-echo "Start cloning the git repositories: tools, python-backend, javascript-frontend"
-cd /home/$THEUSER/$pasta_src
-sudo -u $THEUSER git clone -b ELN_Extractor_Simplify https://jugit.fz-juelich.de/pasta/main.git >> installPASTA2.log  2>&1
-sudo -u $THEUSER git clone -b ELN_Extractor_Simplify https://jugit.fz-juelich.de/pasta/gui.git  >> installPASTA2.log  2>&1
-echo
-
-
-echo "Adopt path and python-path in your environment"
-sudo -u $THEUSER echo "#PASTA changes" >> /home/$THEUSER/.bashrc
-sudo -u $THEUSER echo "export PATH=\$PATH:/home/${THEUSER}/${pasta_src}/main" >> /home/$THEUSER/.bashrc
-sudo -u $THEUSER echo "export PYTHONPATH=\$PYTHONPATH:/home/${THEUSER}/${pasta_src}/main" >> /home/$THEUSER/.bashrc
-echo
-
-
-echo "Install python requirements. This takes a few minutes."
-cd /home/$THEUSER/$pasta_src/main
-sudo -H pip3 install -r requirements.txt           >> installPASTA2.log
-echo
-
-
-echo "Create PASTA configuration file .pasta.json in home directory"
-sudo -u $THEUSER echo "{" > /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"softwareDir\": \"/home/${THEUSER}/${pasta_src}/main\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"userID\": \"${THEUSER}\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"version\": 1," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"extractors\": {}," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"qrPrinter\": {}," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"default\": \"research\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"magicTags\": [\"P1\",\"P2\",\"P3\",\"TODO\",\"WAIT\",\"DONE\"]," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  " >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"links\": {" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    \"research\": {" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "      \"local\": {" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "        \"user\": \"${CDB_USER}\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "        \"password\": \"${CDB_PASSW}\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "        \"database\": \"pasta_tutorial\"," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "        \"path\": \"/home/${THEUSER}/${pasta}\"" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "      }," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "      \"remote\": {}" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    }" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  }," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  " >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  \"tableFormat\": {" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    \"x0\":{\"-label-\":\"Projects\",\"-default-\": [22,6,50,22]}," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    \"measurement\":{\"-default-\": [24,7,23,23,-5,-6,-6,-6]}," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    \"sample\":{\"-default-\": [23,23,23,23,-5]}," >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "    \"procedure\":{\"-default-\": [20,20,20,40]}" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "  }" >> /home/$THEUSER/.pastaELN.json
-sudo -u $THEUSER echo "}" >> /home/$THEUSER/.pastaELN.json
-sudo chown $THEUSER:$THEUSER /home/$THEUSER/.pastaELN.json
 echo
 
 
@@ -165,8 +166,6 @@ echo "Run a short test for 10-20sec?"
 sudo PYTHONPATH=/home/$THEUSER/$pasta_src/main -u $THEUSER python3 Tests/verifyInstallation.py
 echo
 
-
-echo "REMEMBER TO TAKE OUT TEST BRANCH "
 
 echo "Graphical user interface GUI"
 cd /home/$THEUSER/$pasta_src/gui
